% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBFA.R
\name{sBFA_DL}
\alias{sBFA_DL}
\title{Implementation of the data-augmneted Gibbs sampler for the sparse Bayesian factor models with the Dirichlet-Laplace priors}
\usage{
sBFA_DL(
  x,
  k,
  starting,
  priors,
  nmcmc = 10000,
  nburnin = 5000,
  verbose = TRUE,
  nreport = 500
)
}
\arguments{
\item{x}{a matrix containing data}

\item{k}{the number of factors to be fitted.}

\item{starting}{a list with each tag corresponding to a parameter name.
Valid tags are 'Lambda', 'U', 'Sigma', 'Phi', 'tau', 'Psi'.
The value portion of each tag is the parameters' starting values for MCMC.}

\item{priors}{a list with each tag corresponding to a parameter name.
Valid tags are 'Sigma' and 'Phi'.
The value portion of each tag defines the hyperparameters.}

\item{nmcmc}{the number of MCMC iterations.}

\item{nburnin}{the number of burn-in samples.}

\item{verbose}{if TRUE, progress of the sampler is printed to the screen.
Otherwise, nothing is printed to the screen.}

\item{nreport}{the interval to report the MCMC progress.}
}
\value{
a list that contains MCMC samples from the posterior distribution of the sparse Bayesian factor models with the Dirichlet-Laplace priors .
}
\description{
Implementation of the data-augmneted Gibbs sampler for the sparse Bayesian factor models with the Dirichlet-Laplace priors
}
